Move expectation classes to files
