+++ src/Simulator.php
@@ -19,6 +19,10 @@ use Lhsazevedo\Sh4ObjTest\Simulator\Types\U32;
 use Lhsazevedo\Sh4ObjTest\Simulator\Types\U4;
 use Lhsazevedo\Sh4ObjTest\Simulator\Types\U8;
 use Lhsazevedo\Sh4ObjTest\Simulator\Types\UInt;
+use Lhsazevedo\Sh4ObjTest\Test\Expectations\CallExpectation;
+use Lhsazevedo\Sh4ObjTest\Test\Expectations\ReadExpectation;
+use Lhsazevedo\Sh4ObjTest\Test\Expectations\StringWriteExpectation;
+use Lhsazevedo\Sh4ObjTest\Test\Expectations\WriteExpectation;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
