+++ src/Test/Expectations/AbstractExpectation.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Lhsazevedo\Sh4ObjTest\Test\Expectations;
+
+abstract class AbstractExpectation {}
